Refactor for organizational improvements.
