+++ CodeSchool/Docs/DELIVERY_SUMMARY.md
@@ -7,8 +7,8 @@ A comprehensive C# learning project with 2 complete classes, 5 framework classes
 ## Quick Stats
 
 - **Source Code**: 7 C# classes (~850 lines)
-- **Unit Tests**: 110+ tests across 2 test files
-- **Documentation**: 6 comprehensive guides (~2100 lines)
+- **Unit Tests**: 110+ tests across individual test files
+- **Documentation**: 7 comprehensive guides (~2100 lines)
 - **Time to Complete**: 30-40 hours for students
 - **Difficulty**: Beginner to Intermediate
 
@@ -17,36 +17,28 @@ A comprehensive C# learning project with 2 complete classes, 5 framework classes
 ### ✅ Source Code Files
 
 **Foundation Classes (Complete & Documented)**
-- `BankingCS/Transaction.cs` - Immutable transaction model
-- `BankingCS/BankAccount.cs` - Basic account with deposits/withdrawals
+- `../BankingCS/Transaction.cs` - Immutable transaction model
+- `../BankingCS/BankAccount.cs` - Basic account with deposits/withdrawals
 
 **Framework Classes (Ready for Student Implementation)**
-- `BankingCS/CategorizedTransaction.cs` - Inheritance exercise
-- `BankingCS/InterestBearingAccount.cs` - Abstract class exercise
-- `BankingCS/SavingsAccount.cs` - Method override & business logic
-- `BankingCS/CheckingAccount.cs` - Alternative implementations & feature flags
-- `BankingCS/AccountStatement.cs` - LINQ & data analysis
+- `../BankingCS/CategorizedTransaction.cs` - Inheritance exercise
+- `../BankingCS/InterestBearingAccount.cs` - Abstract class exercise
+- `../BankingCS/SavingsAccount.cs` - Method override & business logic
+- `../BankingCS/CheckingAccount.cs` - Alternative implementations & feature flags
+- `../BankingCS/AccountStatement.cs` - LINQ & data analysis
 
 ### ✅ Test Files (110+ Tests)
 
-- `BankingCS.Tests/UnitTest1.cs` - 40+ tests for foundation classes
-- `BankingCS.Tests/StudentFrameworkTests.cs` - 70+ tests for framework classes
-  - CategorizedTransactionTests (7 tests)
-  - SavingsAccountTests (7 tests)
-  - CheckingAccountTests (8 tests)
-  - AccountStatementTests (10+ tests)
+- `../BankingCS.Tests/BankAccountTests.cs` - 40+ tests for foundation classes
+- `../BankingCS.Tests/CategorizedTransactionTests.cs` - 7 tests
+- `../BankingCS.Tests/InterestBearingAccountTests.cs` - 5 tests
+- `../BankingCS.Tests/SavingsAccountTests.cs` - 7 tests
+- `../BankingCS.Tests/CheckingAccountTests.cs` - 8 tests
+- `../BankingCS.Tests/AccountStatementTests.cs` - 10+ tests
 
-### ✅ Documentation (6 Guides)
+### ✅ Documentation (7 Guides)
 
-1. **README.md** (300 lines)
-   - Project overview
-   - Getting started guide
-   - 6-phase learning path
-   - Debugging tips
-   - Testing strategy
-   - Common errors & solutions
-
-2. **BANKING_CURRICULUM.md** (400 lines)
+1. **BANKING_CURRICULUM.md** (400 lines)
    - Complete learning curriculum
    - Learning objectives per phase
    - Key concepts covered
@@ -54,7 +46,7 @@ A comprehensive C# learning project with 2 complete classes, 5 framework classes
    - Real-world connections
    - Assessment rubric
 
-3. **QUICK_REFERENCE.md** (400 lines)
+2. **QUICK_REFERENCE.md** (400 lines)
    - C# syntax reference
    - Properties, methods, constructors
    - Collections & LINQ patterns
@@ -63,7 +55,7 @@ A comprehensive C# learning project with 2 complete classes, 5 framework classes
    - Testing patterns
    - Debugging patterns
 
-4. **INSTRUCTOR_GUIDE.md** (500 lines)
+3. **INSTRUCTOR_GUIDE.md** (500 lines)
    - Complete teaching guide
    - Week-by-week schedule
    - Discussion points for each phase
@@ -72,19 +64,32 @@ A comprehensive C# learning project with 2 complete classes, 5 framework classes
    - Teaching strategies
    - Extension activities
 
-5. **IMPLEMENTATION_SUMMARY.md** (200 lines)
+4. **IMPLEMENTATION_SUMMARY.md** (200 lines)
    - Project overview
    - What has been created
    - How to use the project
    - Success metrics
    - Next steps
 
-6. **FILE_GUIDE.md** (300 lines)
+5. **FILE_GUIDE.md** (300 lines)
    - Index of all files
    - Reading order
    - Quick reference by topic
    - Progress tracking
    - Help resources
+
+6. **DELIVERY_SUMMARY.md** (This file - 300+ lines)
+   - Project overview
+   - Quick statistics
+   - Deliverables checklist
+   - Integration with C# 101
+   - Student outcomes
+
+7. **INDEX.md** (400+ lines)
+   - Navigation hub
+   - Quick start instructions
+   - Concept finder
+   - Troubleshooting guide
 
 ## Key Features
 
@@ -167,8 +172,8 @@ A comprehensive C# learning project with 2 complete classes, 5 framework classes
 5. Refer to QUICK_REFERENCE.md for syntax help
 
 ### For Students
-1. Start with README.md Getting Started
-2. Study Transaction.cs and BankAccount.cs
+1. Start with ../README.md Getting Started
+2. Study ../BankingCS/Transaction.cs and ../BankingCS/BankAccount.cs
 3. Follow the 6-phase learning path
 4. Implement framework classes one by one
 5. Use QUICK_REFERENCE.md for syntax help
@@ -233,16 +238,22 @@ For students who finish early or want more challenge:
 - ✅ AccountStatement.cs (framework)
 
 ### Tests (CodeSchool/BankingCS.Tests/)
-- ✅ UnitTest1.cs (40+ foundation tests)
-- ✅ StudentFrameworkTests.cs (70+ framework tests)
+- ✅ BankAccountTests.cs (40+ foundation tests)
+- ✅ CategorizedTransactionTests.cs (7 tests)
+- ✅ InterestBearingAccountTests.cs (5 tests)
+- ✅ SavingsAccountTests.cs (7 tests)
+- ✅ CheckingAccountTests.cs (8 tests)
+- ✅ AccountStatementTests.cs (10+ tests)
 
-### Documentation (CodeSchool/)
-- ✅ README.md (getting started)
-- ✅ BANKING_CURRICULUM.md (complete curriculum)
-- ✅ QUICK_REFERENCE.md (C# syntax)
-- ✅ INSTRUCTOR_GUIDE.md (teaching guide)
-- ✅ IMPLEMENTATION_SUMMARY.md (overview)
-- ✅ FILE_GUIDE.md (file index)
+### Documentation (CodeSchool/Docs/ and CodeSchool/)
+- ✅ ../README.md (getting started)
+- ✅ Docs/BANKING_CURRICULUM.md (complete curriculum)
+- ✅ Docs/QUICK_REFERENCE.md (C# syntax)
+- ✅ Docs/INSTRUCTOR_GUIDE.md (teaching guide)
+- ✅ Docs/IMPLEMENTATION_SUMMARY.md (overview)
+- ✅ Docs/FILE_GUIDE.md (file index)
+- ✅ Docs/DELIVERY_SUMMARY.md (this file)
+- ✅ Docs/INDEX.md (navigation hub)
 
 ### Project Files
 - ✅ CodeSchool.sln (solution file)
@@ -275,15 +286,15 @@ The project is ready for classroom use when:
 ## Support & Questions
 
 ### For Instructors
-- Review INSTRUCTOR_GUIDE.md for complete teaching plan
-- Check BANKING_CURRICULUM.md for learning objectives
+- Review Docs/INSTRUCTOR_GUIDE.md for complete teaching plan
+- Check Docs/BANKING_CURRICULUM.md for learning objectives
 - Use test failures to guide feedback
-- Reference StudentFrameworkTests.cs to see expectations
+- Reference test files to see expectations
 
 ### For Students
-- Start with README.md Getting Started
+- Start with ../README.md Getting Started
 - Use QUICK_REFERENCE.md for syntax help
-- Study completed classes (Transaction, BankAccount)
+- Study completed classes (../BankingCS/Transaction.cs, ../BankingCS/BankAccount.cs)
 - Read test files to understand what's expected
 - Use TODO comments in framework classes as guidance
 
@@ -292,7 +303,7 @@ The project is ready for classroom use when:
 ```
 Total Lines of Code:           ~850 lines (7 classes)
 Total Lines of Tests:          ~1500 lines (110+ tests)
-Total Lines of Documentation:  ~2100 lines (6 guides)
+Total Lines of Documentation:  ~2100 lines (7 guides)
 ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
 TOTAL PROJECT:                 ~4450 lines
